fix(containerised-app): require Tool Kit version which supports array substitution
We generally try to make our dotcom-tool-kit dependency ranges as broad
as possible as they're a peer dependency and we don't want to limit
version resolution. But for these plugins we need to depend on at least
v4.8.0 as the plugins make use of array substitution of HakoDeploy's
`environments` option, support for which was only added in that version.