fix: ensure tasks exit the command on failure
Many of the tasks added by CP Reliability rethrow errors to ensure that
we catch and explain errors that are outside of our control, e.g.
loading a file that doesn't exist. We wanted these to be Tool Kit errors
so that we get nice error rendering. We also wanted to just throw normal
errors for things like validation errors or missing data.

The issue is that we didn't know about the `exitCode` on Tool Kit error.
This means that any `Error` instances that we wrap will not exit the
command early. This is most notable in the Docker build/deploy commands:
If the Docker build fails then we still move on to try and deploy even
though there's no point. This results in longer build times, unnecessary
tasks being run, and confusing error messages.

We now no longer wrap Tool Kit errors themselves, this allows us to
throw them with the process exit code intact.