fix(circleci-deploy): make deployments depend on build job
When deploying Heroku apps, we’d send the repository to Heroku and it
would run our npm scripts to build and install everything that was
needed for our Node packages into a Docker image. As part of the
migration to AWS, we now need to build this image ourselves in CI. As
the deployment building was done remotely on Heroku’s servers
previously, the `build:ci` command, required for the `test:ci` command,
was run in parallel to the staging deployment. However, now that we need
to build the Docker image in CI, we need those build artefacts before
building the Docker image, so the deployment has a new dependency on the
`build:ci` command.

Let’s update our CircleCI hook config so that the `deploy-staging` job
now depends on the `build` job. This won’t be a breaking change – the
only thing that teams need to do is rerun `npx dotcom-tool-kit
--install` which we don’t consider breaking. This will mean that the
parallelism of the CircleCI tool-kit workflow is reduced, but this is
necessary for AWS builds and Heroku builds are being phased out and will
no longer be supported at some point once the migration is complete.