fix(containerised-app): Revert "feat: update default hako review environment to ft-com-review-eu"
This reverts commit a94b72e67b63eb33764eb617f5bd6f58bbba3fa8.

In reality this should be a breaking change as it means projects will
now need an additional Hako config file for the ft-com-review-eu
environment. Let's revert this for now so that users updating to the
latest version don't encounter CI failures and work out how to support a
migration to a new major version (if necessary) separately.