feat!: update default hako review environment to ft-com-review-eu
This reverts commit 5a55116fe716725a6713d3a05dc60ac423513ca3.

For apps that aren't explicitly setting their review app environment,
this is a breaking change, as those apps will need an ft-com-review-eu/app.yaml
Hako config file to be able to deploy an ephemeral app.

When updating to v0.3 of @dotcom-tool-kit/containerised-app(-with-assets),
you'll either need to explicitly set the `hakoReviewEnvironments` option
to `['ft-com-test-eu']`, or create an ft-com-review-eu/app.yaml config.