feat(containerised-app-with-assets)!: update default hako review environment to ft-com-review-eu
Making a trivial commit as this plugin needs a major version bump after
updating containerised-app.

The migration notes for v0.3.0 are as follows:

For apps that aren't explicitly setting their review app environment,
this is a breaking change, as those apps will need an ft-com-review-eu/app.yaml
Hako config file to be able to deploy an ephemeral app.

When updating to v0.3 of @dotcom-tool-kit/containerised-app(-with-assets),
you'll either need to explicitly set the `hakoReviewEnvironments` option
to `['ft-com-test-eu']`, or create an ft-com-review-eu/app.yaml config.