feat(telemetry): add package to send client side metrics
This package allows us to send metrics from Tool Kit invocations to a
backend server, currently planned to be an instance of Vector running in
AWS. We can use these metrics to get visibility for how people use Tool
Kit, and where things may be going wrong. For now, I've only added one
metric, which tracks what Tool Kit tasks are run and whether they
completed successfully.

The telemetry package uses a child process to allow us to send the
metrics asynchronously and in the background once the main Tool Kit
process exits, meaning there is minimal impact on the time it takes for
Tool Kit commands to run on developer machines.

Children of the metrics class can be created with arbitrary attributes
overridden, and these classes are passed around between functions and
stored as fields in Task and Hook classes, allowing important metadata,
such as whether we're in CI or the name of the task currently running,
to always be included in recorded events and scoped logically.