chore: fix release-please by marking last release-please PR
We added a commit in the Release Please PR to bump the peer dependencies
of plugins to use the new version of the core Tool Kit package. This
commit only made changes for the `plugins` packages, not packages in
`core` and `lib`. However, Release Please tags all package releases with
that latest commit. This means that it runs into issues later on when
trying to find the last commit that a given package was released for, as
it can see there's a release tagged for that latter commit, but filters
out that commit as a candidate for the packages outside of `plugins` as
nothing changed in their paths. This meant that it was effectively
considering all changes since the release before last for those
packages.

Let's work around this by explicitly setting when the last release was,
which will cause Release Please to not look at any commits before this
time. The documentation warns that this SHA is never ignored so should
be removed 'once a good release PR is merged', but it seems to be that
we would run into the same filtering issue until _all_ packages have
another release since the last one. Looking at the code, it doesn't seem
that leaving it around will cause any adverse effects as it will only
cause the commit search to end once this commit has been reached, and
won't mean that every subsequent release will be trying to release all
commits since the given SHA.