feat(circleci): add option to specify full git checkout
CircleCI has recently added support for blobless git checkouts, and is
starting to roll out a change that makes them the default checkout
method. Blobless checkouts will download the state of the git repository
at the current commit but won't download all the git 'blobs' that
contain information on file history. This is fine for most projects, but
can cause issues if a project's script call out to something like `git
diff` for instance. git will require the blobs to be able to the diff
the file with its contents from a previous commit so will then run a
`git fetch` to get the relevant blobs, but this will fail in most
CircleCI jobs as they don't have GitHub's public keys installed to allow
git to verify GitHub's ssh host.

Let's allow projects to override the checkout method so that they can
avoid these verification issues if their CI jobs call out to git.